Cross-compile

+++ package.ps1
@@ -2,7 +2,11 @@
 # Copyright (c) 2024 Roger Brown.
 # Licensed under the MIT License.
 
-Param($Version, $Maintainer)
+Param(
+	$Version,
+	$Maintainer,
+	$Release = '1.ubuntu'
+)
 
 $ErrorActionPreference = [System.Management.Automation.ActionPreference]::Stop
 $WorkDir = 'github-PowerShell'
@@ -38,16 +42,6 @@ if (-not $Maintainer)
 	}
 }
 
-$Architecture = ( sh -c "dpkg --print-architecture" ).Trim()
-
-$Architecture
-
-$RID = [System.Runtime.InteropServices.RuntimeInformation]::RuntimeIdentifier
-
-$Arch = ($RID.Split('-'))[-1]
-
-$Arch
-
 if (-not ( Test-Path $WorkDir ))
 {
 	sh -c "git clone https://github.com/PowerShell/PowerShell.git $WorkDir --single-branch --branch $ReleaseTag"
@@ -71,7 +65,11 @@ if (-not ( Test-Path $WorkDir ))
 		$xml.Save("$PWD/$ProjectFile")
 		Import-Module ./build.psm1
 		Start-PSBootstrap
-		Start-PSBuild -Configuration Release -Clean -ReleaseTag $ReleaseTag -Runtime "linux-$Arch"
+
+		foreach ($Arch in 'arm', 'arm64', 'x64')
+		{
+			Start-PSBuild -Configuration Release -ReleaseTag $ReleaseTag -Runtime "linux-$Arch"
+		}
 	}
 	finally
 	{
@@ -88,53 +86,54 @@ if (-not (Test-Path $OriginalFile))
 	Invoke-WebRequest -Uri $Uri -OutFile $OriginalFile
 }
 
-$Release = '1.ubuntu'
-
-$OutputFile = "powershell_$Version-$Release`_$Architecture.deb"
-
-if (Test-Path $OutputFile)
-{
-	Remove-Item $OutputFile
-}
-
 $ReleaseDir = "$WorkDir/$ProjectDir/bin/Release"
 
-$Null = New-Item 'control',
-		'data',
-		'data/usr',
-		'data/usr/bin',
-		'data/opt',
-		'data/opt/microsoft',
-		'data/opt/microsoft/powershell' -ItemType Directory
+Get-ChildItem $ReleaseDir -Name 'publish' -Directory -Recurse | ForEach-Object {
+	$Configuration = $_
 
-Get-ChildItem $ReleaseDir -Directory -Name 'publish' -Recurse | ForEach-Object {
-	$SrcDir = Join-Path $ReleaseDir $_
-	sh -c "find $SrcDir -type l | xargs --no-run-if-empty rm"
-	Copy-Item $SrcDir 'data/opt/microsoft/powershell/7' -Recurse
-}
+	$ConfigurationElements = $Configuration.Split('/')
 
-$DotnetRuntime = Get-ChildItem $ReleaseDir -Directory -Name 'publish' -Recurse | ForEach-Object {
-	$_.Split('/') | ForEach-Object {
-		if ($_.StartsWith('net'))
-		{
-			$_.Replace('net','dotnet-runtime-')
+	if ($ConfigurationElements.Length -eq 3)
+	{
+		$DotnetRuntime = $ConfigurationElements[0].Replace('net','dotnet-runtime-')
+
+		$Architecture = $ConfigurationElements[1].Split('-')[-1]
+
+		$ArchMap = @{
+			'arm' = 'armhf'
+			'arm64' = 'arm64'
+			'x64' = 'amd64'
 		}
-	}
-}
 
-$DotnetRuntime
+		$Architecture = $ArchMap.$Architecture
 
-sh -e -c "cd data ; ar p ../$OriginalFile data.tar.gz | tar xvfz - ./usr/local/share/man/man1/pwsh.1.gz"
-		
-if ($LastExitCode)
-{
-	exit $LastExitCode
-}
+		$OutputFile = "powershell_$Version-$Release`_$Architecture.deb"
 
-$Size = sh -c 'du -sk data | while read A B; do echo $A; done' 
+		if (Test-Path $OutputFile)
+		{
+			Remove-Item $OutputFile
+		}
 
-$Maintainer
-$Size
+		$Null = New-Item 'control',
+			'data',
+			'data/usr',
+			'data/usr/bin',
+			'data/opt',
+			'data/opt/microsoft',
+			'data/opt/microsoft/powershell' -ItemType Directory
+
+		$SrcDir = Join-Path $ReleaseDir $Configuration
+		sh -c "find $SrcDir -type l | xargs --no-run-if-empty rm"
+		Copy-Item $SrcDir 'data/opt/microsoft/powershell/7' -Recurse
+
+		sh -e -c "cd data ; ar p ../$OriginalFile data.tar.gz | tar xfz - ./usr/local/share/man/man1/pwsh.1.gz ./usr/share/doc/powershell/changelog.gz"
+
+		if ($LastExitCode)
+		{
+			exit $LastExitCode
+		}
+
+		$Size = sh -c 'du -sk data | while read A B; do echo $A; done'
 
 @"
 Package: powershell
@@ -177,19 +176,21 @@ esac
 
 '2.0' | Set-Content 'debian-binary'
 
-foreach ($cmd in 'find data -type f | xargs chmod -x',
-		'chmod ugo+x data/opt/microsoft/powershell/7/pwsh control/postinst control/postrm',
-		'ln -s /opt/microsoft/powershell/7/pwsh data/usr/bin/pwsh',
-		'cd data ; find * -type f -print0 | xargs -r0 md5sum > ../control/md5sum',
-		'cd data ; tar  --owner=0 --group=0 --create --xz --file ../data.tar.xz ./*',
-		'cd control ; tar  --owner=0 --group=0 --create --xz --file ../control.tar.xz ./*',
-		"ar r $OutputFile debian-binary control.tar.xz data.tar.xz",
-		'rm -rf debian-binary control.tar.xz data.tar.xz data control')
-{
-	sh -e -c $cmd
+		foreach ($cmd in 'find data -type f | xargs chmod -x',
+			'chmod ugo+x data/opt/microsoft/powershell/7/pwsh control/postinst control/postrm',
+			'ln -s /opt/microsoft/powershell/7/pwsh data/usr/bin/pwsh',
+			'cd data ; find * -type f -print0 | xargs -r0 md5sum > ../control/md5sum',
+			'cd data ; tar  --owner=0 --group=0 --create --xz --file ../data.tar.xz ./*',
+			'cd control ; tar  --owner=0 --group=0 --create --xz --file ../control.tar.xz ./*',
+			"ar r $OutputFile debian-binary control.tar.xz data.tar.xz",
+			'rm -rf debian-binary control.tar.xz data.tar.xz data control')
+		{
+			sh -e -c $cmd
 
-	if ($LastExitCode)
-	{
-		exit $LastExitCode
+			if ($LastExitCode)
+			{
+				exit $LastExitCode
+			}
+		}
 	}
 }
